Hand-Made Ammo Part 1
What's added:
Tarnished Casings and Shotgun Shells either from age in the zone or poor craftsmanship.

What's it do:
Currently for part 1, all of the ammo is there, as well as: the icons, boxes, items, and projectiles.
Loottables for stashes were touched just slightly to add in the new ammo with the original ammo's chance of 10, while the rest of the regular ammo has been dropped down to 7, to signify that better rounds are harder to find with no true way to import them into the zone.

What's next:
Well. Part 2 will hopefully have some tweaks to the bullets based on suggestions. I'll add them for sale in the SidorVends. And as well, I need to figure out how to make using old rounds take two durability away from a gun instead of the one for normal rounds.