Adds the new melee sprites into stalker and starts converting the old ones
Too little time to answer all my questions when it's 1am, will work on the rest of the sprites and code.

Currently need to find out the meaning of some of the statements and find the sources/instances of what I need to change. Will also add new items depending on current itemos availablos.