Yantar Map Updated + Minor sprite/code tweaks
Yantar has been expanded to the full map, still needs interiors, more dungeons, more broken down vehicles, etc

Changed it so when mutants die it doesnt repeat its name for the last death message "the snork the snork keels over and dies" to "the snork keels over and dies" example given

Re-colored some of the jungle assets so that they're not so bright green, to blend in more with other assets.

Changed what Sidorovich says from the search and kill stalkers with a bad rep with 'Dont go wandering alone and die, you have much more to buy from me!'

Changed the teleporter in Radar to go to the appropriate new place in Yantar.