Army Warehouse Fix + Lotsa Resprites + Sprites
Added some sprites, resprote some stuff!! More assets!! Check the diffs to see whats added.

Some of these sprites are not coded in yet, but there will be a seperate PR which will add them to lootspawner as well as code them as actualy items. Pushing now because this map fix needs to be merged ASAP to fix it. When the previous mapper made the custom description for the bloodsucker in captivity, they forgot to put commas around the desc and eng_desc.

Updated the Freedom base very slightly and a little bit elsewhere around Army Warehouse.

Removed text that says 'NO ACCESS' when examining