Okay. New fraction Renegate helmet
To be honest, it's not very authentic. I wanted to make a dumpy helmet out of a clear-sky helmet and a helmet, but it came out some box with a dead visor from the "super soldier" kit, which doesn't really resemble the bum Renegade. However, all in all, better than nothing? It's, to me at least, prettier than the tactical helmets.

And yes, it's done, all that's left is the merge. All checked, zero errors, all bought in SIDORMAT'e.