Bugfixes + Mapper Icons
Fixed loot spawns looking literally like a loot spawn marker on a few places of the map

Fixed the boar hoof icon_state, so it should appear now with its sprite

Fixed the vodka bottles disappearing and not giving broken sprites.

Fixed the icon state path for the twohanded old axe

Fixed the icon state path for both variants of the broken bear detector

Fixed the icon name for fermenting barrels, so it should appear now

Made a bunch of new icons for loot, I'll add more to it but this is a good start. It can only be seen by mappers, so now stalker 13 mappers can identify what loot is around at an eyes glance.

Went around and fixed all the teleporters I could find, let me know if there are any I missed!

Added 3 unused teleporters from the ancient iWuna age

Spruced up a few thing on the map

There was a bug in the floor code for vertical wooden tiles to select for random from 1 out of 70, but we only have 25 different, causing it to select a generic floor. All fixed up.

Added Nagant revolver to the buy/sell list on sidoramat. Also nerfed the damage so it wasnt as powerful as a 50 Cal!

Changed the initial stock of items from 5 to 10. There is a function that restocks between every 15 to 30 minutes if someone clears it out.

Changed the total quantity of items in IFAK to 7. Huge nerf, but it didnt need to have that much in the first place, and if it did have 16 slots it would be overpowered.

Changed the blowout siren from hitting around 0 dB (loud) to -6 (still loud but not crazy loud) More sound level balancing will be done later.

https://i.imgur.com/1jSSmzI.png
https://i.imgur.com/2PXsUCh.png
https://i.imgur.com/mK540d4.png
https://i.imgur.com/ipMOfRO.png
https://i.imgur.com/vektwhc.png
https://i.imgur.com/gRGXMi7.png